Move missed functions to gcp_config.go
